+++ src/main/java/nadeuli/entity/ExpenseItem.java
@@ -55,7 +55,7 @@ public class ExpenseItem extends BaseTimeEntity {
     private String content;
 
     @Column(name = "expense", columnDefinition = "INT UNSIGNED not null")
-    private Integer expense;
+    private Long expense;
 
 
 //    // 생성자
@@ -70,7 +70,7 @@ public class ExpenseItem extends BaseTimeEntity {
 //    }
 
     // static factory method
-    public static ExpenseItem of (ExpenseBook ebid, ItineraryEvent ieid, Traveler payer, String content, Integer expense) {
+    public static ExpenseItem of (ExpenseBook ebid, ItineraryEvent ieid, Traveler payer, String content, Long expense) {
         return new ExpenseItem(null, ebid, ieid, payer, content, expense);
     }
